\name{assignEvents}
\alias{events}
\alias{EVENTS}
\alias{extractEvents}
\alias{loadEvents}

\title{
assign, extract, and load events (transits) from the RAW_YYYY tables to the BETA_EVENTSYYYY tables
}
\description{
clusters RAW_YYYY data, predicts direction based on 2013 chick-feeding calibration data, and saves the result to BETA_EVENTSYYYY.
}
\usage{
loadEvents(year)
}
