% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EVENTS.R
\name{birdIDs}
\alias{birdIDs}
\title{retrieve all bird ringnumbers/cb code/transponder/genetic sex combinations
from the database}
\usage{
birdIDs()
}
\value{
a dataframe with variables birdID,transponder,cb,sex. each distinct
  combination has one line, so if a bird gets retraspondered it will appear
  more than once in this dataframe. the fieldteam gets birdID 'fieldteam'.
}
\description{
query  BTatWESTERHOLZ.ADULTS, BTatWESTERHOLZ.CHICKS,
FIELD_BTatWESTERHOLZ.ADULTS, and SNBatWESTERHOLZ2.test_transponders (a view
that trawls all authors tables) to find all existing
ringnumber/transponder/colourbandcombo/genetic sex combinations.
}

