% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getData.R
\name{boxon_data}
\alias{boxon_data}
\title{identify intervals where the box is most likely 'on'}
\usage{
boxon_data(con, startd=Sys.Date()-31,endd=Sys.Date(), bx= c(1:277))
}
\arguments{
\item{con}{an open connection (to SNBatWESTERHOLZ2, but also works for
SFatWESTERHOLZ)}

\item{startd}{starting date of the period of interest}

\item{endd}{end date of the period of interest}

\item{bx}{box (or vector of boxnumbers) for which to find 'on' periods}
}
\value{
a dataframe with, for each downloaded file, the boxnumber, download
  date, start date (truncated to the given startd), the last datetime
  (truncated to the given endd), the upload status, the total number of days
  covered by this chunk of time, and the total number of 'on' days in this
  chunk of time. Currently, this function is used to plot the background
  rectangles for the actograms.
}
\description{
within a given period of time, find all intervals where the
  specified box(es) were, according to themselves, 'on'. As in, fully
  functional.
}
\section{NB}{
 if onDays is larger than totalDays, the file has an RTC problem.
}

\section{POINT OF IMPROVEMENT}{
 This function currently only uses data
  available in the file_status table, because that is much faster than
  trawling RAW_YYYY. However, this approach means that for instance
  intermittent malfunctioning (where the box switches to garble halfway
  through the file, but switches back to normal behaviour later on) is
  missed, and that the function errs on the side of type 2 error. I think
  this function should be as conservative as possible so that you can trust
  that IF a box displays as 'ON' it fucking well IS on. So i think the
  function should either query RAW_YYYY, or flag RTC-problem files (because
  boxes that are backwards-jumping will still produce valid datetimes within
  the start and end limits of the file)
}

