% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EVENTS.R
\name{ct}
\alias{ct}
\title{conditional inference tree used to assign direction}
\usage{
data(ct)
}
\description{
decision tree, based on the 2013 calibrations, used to predict direction for
previously extracted and summarised events.
}
\details{
\itemize{ \item the tree was trained using party::ctree on the first
  ten visits for each video in the 2013 calibrations (351 visits total). In
  the tree data, LB = -1 means only the outer barrier is firing; 0 means both
  barriers are firing, and 1 means only the inner barrier fires. \item as the
  tree is based on chick-feeding data, predictions on direction outside the
  chick-feeding period are inaccurate. Because of the nature of the feeding
  visits, most transits outside the breeding season are classified as 'out'.}
}
\section{tree}{
 plot(ct)
}
\seealso{
\code{\link[party]{ctree}},\href{http://scidb.orn.mpg.de/scidbwiki/westerholz/doku.php?id=smart_nest_box_project:calibrations:2013}{snb
calibrations 2013}
}

