% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EVENTS.R
\name{extractEvents}
\alias{extractEvents}
\title{extract and summarise events}
\usage{
extractEvents(d)
}
\arguments{
\item{d}{a dataframe as returned by assignEvents()}
}
\value{
a dataframe with, for each valid event, a number of parameters that
 can be used to assign direction (which happens in loadEvents())
}
\description{
extract and summarise events
}
\details{
various parameters that in the calibration data differ between 'ins'
 and 'outs' are extracted for each event: \itemize{\item the duration of the
 event (in s) of each visit (room for improvement: could be in ms). Shorter
 for 'in'. \item the number of lines with no transponderreading before the
 first transponderreading. higher for 'ins' for birds that hang. \item the
 first and last LB reading of the event. \item the predicted direction based
 on the first and last LB reading. \item whether there is an actual transit
 (both LB's firing) occurring during the event. \item the number of rows in
 the event (disregarding bout_length, so not the actual number of lines in
 the datafile). Less for 'ins'. \item the average bout_length during the
 event. higher for 'ins'. Higher variance in bout_length for unclear visits
 (not currently considered or calculated; room for improvement). \item the
 sum of all boutlengths in the event. This is the actual number of lines in
 the raw datafile. \item starting time, number of lines, sum of all
 boutlenghts, and direction based on LB readings for the previous event per
 transponder. \item the ratio between the number of lines in the current
 event and the number of lines in the previous event for that transponder.}
 Only events that contain a transit or a transponderreading (or both) are
 returned.
}
\seealso{
\code{\link{assignEvents}} for assigning an event ID to the raw
 data, \code{\link{loadEvents}} for the wrapper function that identifies,
 extracts, predicts direction, and loads events into BETA_EventsYYYY.
}

