

\name{extractVars}
\alias{extractVars}
\title{extracts snb variables from read.raw output}
\description{extracts datetime,boutlength,bv,transp,pir,lb from variable d$V1 in the output generated by read.raw()}
\usage{extractVars(d)}
\arguments{
	\item{d}{dataframe generated by read.raw(path)}
	}

\note{
d should contain the following data on the following positions:
\itemize{
\item{column 1}{full record}
\item{column 2}{bout length}
\item{column 3}{file status}
}
column names are not important.
}
\seealso{\code{\link{loadSingle}, \link{loadDbase}, \link{readSingle}}}

\examples{

# extract records
path = "\\\\\\\\FILES\\\\raw_data_kemp\\\\FIELD\\\\Westerholz\\\\breeding_2008\\\\Smart_Nest_Box\\\\2008.05.21\\\\276\\\\BOX001.TXT"
d = read.raw(path)
extractVars(d)
}


\author{AR 04.05.2011}

