% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EVENTS.R
\name{loadEvents}
\alias{loadEvents}
\title{assign, summarise, and upload events}
\usage{
loadEvents(year_ = substring(Sys.Date(), 1, 4))
}
\arguments{
\item{year_}{year for which to find unloaded events and assign, extract, and
load these.}
}
\value{
TRUE if the data get processed and loaded correctly.
}
\description{
\code{loadEvents} is the wrapper around
 \code{\link{assignEvents}}, which assigns a boutID to the raw data, and
 \code{\link{extractEvents}}, which extracts and summarises eligible events.
 \code{loadEvents} assigns event ids to previously unprocessed raw data,
 extracts and summarises events that contain an actual transit (both LB's
 fireing) or a transponderreading or both, assigns direction to these events
 using the `ct` inference tree object in the package to assign direction to
 each event, and loads to BETA_EventsYYYY.
}
\details{
pipeline: \enumerate{\item looks for fileids that are in
 SNBatWESTERHOLZ2.file_status but not in BETA_EventsYYYY.(room for
 improvement: this way, files with no valid events are processed every time
 because the fileid will never show up in any events table). \item for each
 fileid, queries RAW_YYYY for all data from this file. \item matches
 transponder to bird/fieldteam ID. \item saves one record for each distinct
 transponder with invalid datetimes. \item assigns events using
 \code{\link{assignEvents}}. \item summarises and extracts valid events using
 \code{\link{extractEvents}}. \item predicts event direction using
 \code{\link{ct}}. \item saves events to BETA_EventsYYYY.} If you want to
 re-assign events for a certain file, delete all occurences of that fileid
 from BETA_EventsYYYY.
}
\seealso{
\code{\link{extractEvents}} for extracting and summarising events,
 \code{\link{assignEvents}} for assigning an event ID to the raw data.,
 \code{\link{ct}} for a description of the inference tree used to predict
 direction.
}

