% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EVENTS.R
\name{loadSFEvents}
\alias{loadSFEvents}
\title{identify, extract, and upload smartfeeder visits}
\usage{
loadSFEvents(cutoff = 2)
}
\arguments{
\item{cutoff}{if the time in seconds elapsed between two readings is larger
than the cutoff value, data gets assigned to a new visit.}
}
\value{
TRUE if data gets loaded successfully.
}
\description{
\code{loadSFEvents} assigns, extracts, and loads events from smartfeederdata
from SFatWESTERHOLZ.RAW_YYYY. Visit data gets loaded into
SFatWESTERHOLZ.EventsYYYY.
}
\details{
pipeline similar to \code{\link{loadEvents}}, but event assignment
  and summarisation is simplified and no direction prediction takes place.
  Event assignment is based on time elapsed since previous reading only.
}

