% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{writeload}
\alias{writeload}
\title{write to tempfile & load to db}
\usage{
writeload((d, fname = gsub('\\\\','/',tempfile()), con, db, tb, ignore
  = FALSE, charset = 'latin1')
}
\arguments{
\item{d}{a dataframe for upload to table db.tb, note that function does not
check variable names.}

\item{fname}{tempfile used to save data.}

\item{con}{an open connection to db.tb}

\item{db}{database where table tb is}

\item{tb}{tablename to write data to}

\item{ignore}{should conflicts be ignored}

\item{charset}{character set dataframe d is in. Loading special characters
  from one character set to a db.tb in a different character set leads to
  problems.}
}
\description{
write to tempfile & load to db
}
\details{
this function roughly does what \code{\link{dbWriteTable}} does,
  except you can specify IGNORE which means you can load data into a table
  with a UNIQUE index and thus avoid uploading duplicates, which is useful
  for BETA_Events (because CF files my contain duplicate data) and
  META_SnbStatus.
}

